+++ mdadash/backend/tests/data/files.py
@@ -10,6 +10,7 @@ Use as ::
 
 __all__ = [
     "TPR",
+    "TRR",
     "XTC",
 ]
 
@@ -18,4 +19,5 @@ import importlib.resources
 data_directory = importlib.resources.files("mdadash") / "backend" / "tests" / "data"
 
 TPR = data_directory / "adk_oplsaa.tpr"
+TRR = data_directory / "adk_oplsaa.trr"
 XTC = data_directory / "adk_oplsaa.xtc"